requirements.txt
fastapi
uvicorn
httpx
aiosmtplib
python-dotenv
asyncpg